jest: mainRequest and staticAssets ported
